+++ invest/redirects.py
@@ -1,0 +1,49 @@
+REDIRECTS = {
+    'high-potential-opportunities/rail-infrastructure': '/international/content/invest/high-potential-opportunities/rail-infrastructure/',  # NOQA
+    'high-potential-opportunities/food-production': '/international/content/invest/high-potential-opportunities/food-production/',  # NOQA
+    'high-potential-opportunities/lightweight-structures': '/international/content/invest/high-potential-opportunities/lightweight-structures/',  # NOQA
+    'high-potential-opportunities/rail-infrastructure/contact': '/international/content/invest/high-potential-opportunities/rail-infrastructure/contact/',  # NOQA
+    'high-potential-opportunities/food-production/contact': '/international/content/invest/high-potential-opportunities/food-production/contact/',  # NOQA
+    'high-potential-opportunities/lightweight-structures/contact': '/international/content/invest/high-potential-opportunities/lightweight-structures/contact/',  # NOQA
+    # translated urls
+    'contact': '/international/invest/contact/',
+    'uk-regions/midlands': '/international/content/invest/uk-regions/midlands/',
+    'uk-regions/northern-ireland': '/international/content/invest/uk-regions/northern-ireland/',
+    'uk-regions/south-england': '/international/content/invest/uk-regions/south-england/',
+    'uk-regions/north-england': '/international/content/invest/uk-regions/north-england/',
+    'uk-regions/london': '/international/content/invest/uk-regions/london/',
+    'uk-regions/wales': '/international/content/invest/uk-regions/wales',
+    'uk-regions/scotland': '/international/content/invest/uk-regions/scotland/',
+    'industries': '/international/content/industries/',
+    'industries/health-and-life-sciences': '/international/content/industries/health-and-life-sciences/',
+    'industries/pharmaceutical-manufacturing': '/international/content/industries/health-and-life-sciences/',
+    'industries/medical-technology': '/international/content/industries/health-and-life-sciences/',
+    'industries/creative-industries': 'international/content/industries/creative-industries/',
+    'industries/digital-media': '/international/content/industries/creative-industries/',
+    'industries/creative-content-and-production': '/international/content/industries/creative-industries/',
+    'industries/food-and-drink': '/international/content/industries/food-and-drink/',
+    'industries/meat-dairy-and-polutry': '/international/content/industries/food-and-drink/',
+    'industries/food-services-and-catering': '/international/content/industries/food-and-drink/',
+    'industries/free-from-foods': '/international/content/industries/food-and-drink/',
+    'industries/financial-services': '/international/content/industries/financial-services/',
+    'industries/asset-management': '/international/content/industries/financial-services/',
+    'industries/financial-technology': '/international/content/industries/financial-services/',
+    'industries/aerospace': '/international/content/industries/aerospace/',
+    'industries/agri-tech': '/international/content/industries/agricultural-technology/',
+    'industries/retail': '/international/content/industries/retail/',
+    'industries/chemicals': '/international/content/invest/industries/chemicals',
+    'industries/technology': '/international/content/industries/technology/',
+    'industries/data-analytics': '/international/content/industries/technology/',
+    'industries/capital-investment': '/international/content/invest/capital-invest',
+    'industries/automotive': '/international/content/industries/automotive/',
+    'industries/motorsport': '/international/content/industries/automotive/',
+    'industries/automotive-research-and-development': '/international/content/industries/automotive/',
+    'industries/automotive-supply-chain': '/international/content/industries/automotive/',
+    'energy': '/international/content/industries/',
+    'industries/offshore-wind-energy': '/international/content/industries/renewable-energy/',
+    'industries/oil-and-gas': '/international/content/invest/industries/oil-and-gas',
+    'industries/electrical-networks': '/international/content/industries/',
+    'industries/nuclear-energy': '/international/content/invest/industries/nuclear-energy',
+    'industries/energy-from-waste-market': '/international/content/industries/renewable-energy/',
+    'industries/advanced-manufacturing': '/international/content/industries/advanced-manufacturing/'
+}